Tests/LibWeb/Text/expected/WebAudio/compressor-node-channel-cound-mode.txt

clamped-max
explicit
NotSupportedError
clamped-max
explicit
NotSupportedError
